refactor: extract donut chart generation into reusable module
Extract `generateTopCountryDonut` function from statistics report into a generic `getDonutChart` utility in `reports/charts/donut.chart.ts`. The new implementation accepts configurable entries and legend position, making it reusable across different report types. Update statistics report to use the new chart module and reorganize exports to include charts and sections.
