+++ report-pdf/src/reports/charts/index.ts
@@ -1,0 +1,1 @@
+export * from './donut.chart';
