+++ report-pdf/src/reports/index.ts
@@ -5,5 +5,5 @@ export * from './employment-letter-by-id.report';
 export * from './employment-letter.report';
 export * from './order-by-id.report';
 export * from './statistics.report';
-export * from './sections/footer.section';
-export * from './sections/header.section';
+export * from './charts';
+export * from './sections';
